Adding Supports for Records (Groovy) + Passing Record Test.
