+++ jr-objects/src/main/java/com/fasterxml/jackson/jr/ob/impl/BeanPropertyIntrospector.java
@@ -4,6 +4,7 @@ import java.lang.reflect.Constructor;
 import java.lang.reflect.Field;
 import java.lang.reflect.Method;
 import java.lang.reflect.Modifier;
+import java.util.Arrays;
 import java.util.Map;
 import java.util.TreeMap;
 
@@ -141,6 +142,13 @@ public class BeanPropertyIntrospector
                         // May or may not be used, but collect for now all the same:
                         name = decap(name.substring(2));
                         _propFrom(props, name).withIsGetter(m);
+                    }
+                } else {
+                    // This will allow getters with field name as their getters, like the ones generated by Groovy
+                    boolean isDirectNameGetter = Arrays.stream(currType.getDeclaredFields()).map(Field::getName).anyMatch(f -> f.equals(m.getName()));
+                    if (isDirectNameGetter) {
+                        name = decap(name);
+                        _propFrom(props, name).withGetter(m);
                     }
                 }
             } else if (argTypes.length == 1) { // setter?
